Validating fapgans

+++ FapgansControleBot/Repository/CreditRepository/i_credit_repository.py
@@ -3,7 +3,7 @@ from FapgansControleBot.Repository.repository import Repository
 
 
 class ICreditRepository(Repository[Credit]):
-    def find_credit_by_price(self, price: int):
+    def find_credit_by_price(self, price: float):
         raise NotImplementedError
 
     def active_gans_periods(self):
